refactor: organize Timer components into directories for better structure
