+++ src/components/Timer/TimerBackground.tsx
