+++ src/components/Timer/TimerControls.tsx
