+++ src/components/Timer/TimerItem.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useRef, useState } from 'react';
 import { Trash2, RotateCcw, Pencil } from 'lucide-react';
-import { Timer } from '../types/timer';
-import { formatTime } from '../utils/time';
-import { useTimerStore } from '../store/useTimerStore';
+import { Timer } from '../../types/timer';
+import { formatTime } from '../../utils/time';
+import { useTimerStore } from '../../store/useTimerStore';
 import { toast } from 'sonner';
-import { EditTimerModal } from './EditTimerModal';
-import { TimerAudio } from '../utils/audio';
+import { EditTimerModal } from './Modal/EditTimerModal';
+import { TimerAudio } from '../../utils/audio';
 import { TimerControls } from './TimerControls';
 import { TimerProgress } from './TimerProgress';
