+++ src/components/Timer/TimerList.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { TimerItem } from './TimerItem';
-import { useTimerStore } from '../store/useTimerStore';
+import { useTimerStore } from '../../store/useTimerStore';
 import { EmptyState } from './EmptyState';
 
 export const TimerList: React.FC = () => {
