+++ src/components/Timer/TimerProgress.tsx
